Add Markdown Link Checker (#28)
# Description

This change introduces a new workflow step to check Markdown links in the repository. A new configuration file `.linkspector.yml` is added to specify the directories to scan and the acceptable status codes for live links. The `code-quality.yml` workflow is updated to include a new job called `check-markdown-links`, which uses the UmbrellaDocs/action-linkspector action to perform the link checking. 

fixes #27