Improve SonarCloud Configuration and Coverage Reporting (#50)
# Description

This change enhances the SonarCloud integration in the CI/CD pipeline
and adjusts the SonarCloud configuration. The following modifications
were made:

1. Added a step in the GitHub Actions workflow to override the coverage
source path for SonarCloud. This ensures that the correct file paths are
used when reporting code coverage.

2. Updated the `sonar-project.properties` file to exclude test files
from coverage analysis and test execution. This helps to focus the
analysis on the main codebase and provides more accurate metrics.

These changes aim to improve the accuracy and relevance of the
SonarCloud scan results for the project.

fixes #48