Enhance Pull Request Labelling Workflow (#65)
# Description

This change updates the Pull Request Labeller workflow to enhance labelilng functionality. The workflow now uses a specific version of the labeler action (v5.0.0) and corrects the configuration file path to 'labeller.yml'. Additionally, it introduces a new step to add size labels to pull requests using the size-label-action.

fixes #63