Add Pyproject Checks (#133)
# Pull Request

## Description

This pull request includes changes to the GitHub workflow and the
`Justfile` to add validation for the `pyproject.toml` file. The most
important changes include adding a new step in the GitHub workflow to
validate the `pyproject.toml` file and defining a new command in the
`Justfile` to perform this validation.

Changes to GitHub workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R66-R67):
Added a new step named "Validates Pyproject" to run the `just
pyproject-check` command.

Changes to `Justfile`:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR13-R16):
Added a new command `pyproject-check` to validate the `pyproject.toml`
file using `poetry check`.