Schedule Code Checks to Run Daily (#136)
# Pull Request

## Description

This pull request includes a small but important change to the
`.github/workflows/code-checks.yml` file. The change updates the
workflow name and adds a schedule trigger for daily runs.

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L1-R9):
Renamed the workflow from "Code Quality" to "Code Checks" and added a
cron schedule to run the workflow daily at midnight.