Add HackerRank Arrays Challenge (#142)
# Pull Request

## Description

This pull request introduces a new function to reverse an array and
updates the `pyproject.toml` file to ignore additional linting rules.
The most important changes include adding a new function to reverse an
array and updating the linting rules to ignore certain checks.

New function implementation:

*
[`challenges/HackerRank/DataStructures/Arrays/challenge.py`](diffhunk://#diff-f5c83ad61bb06c1c9f9d2c1dfdff40124f92ad8c21ae6b6d89529914fc3a5dd3R1-R29):
Added a new function `reverseArray` to reverse an array of integers and
updated the script to read input, call the function, and write the
output to a file.

Linting rules updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711R33-R36):
Added new rules to the `ignore` list to ignore missing docstrings in
public functions and other specific linting checks.


https://www.hackerrank.com/challenges/arrays-ds/problem?isFullScreen=true#