Fix Security Policy Links (#146)
# Pull Request

## Description

This pull request includes updates to the `docs/SECURITY.md` file to
correct the repository URLs for reporting vulnerabilities and for
accessing security advisories.

Documentation updates:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9):
Updated the URL in the "To report a vulnerability" section to point to
the correct repository security tab.
*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L23-R23):
Updated the URL in the "Security Updates" section to direct users to the
correct repository's security advisories section.