Update Dependency Review (#152)
# Pull Request

## Description

This pull request includes changes to the GitHub Actions workflow
configuration to enhance the Dependency Review job. The most important
changes are:

Workflow configuration updates:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR41-R42):
Added `pull-requests: write` permission to the Dependency Review job to
allow it to write comments on pull requests.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR51-R52):
Configured the Dependency Review action to comment a summary in the pull
request on failure.