Refactor Git Hooks (#156)
# Pull Request

## Description

This pull request includes several updates to pre-commit hooks,
dependency versions, and the addition of new checks for Ruff issues. The
changes aim to improve the robustness of pre-commit checks and ensure
the project dependencies are up to date.

### Pre-commit hooks improvements:

*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL2-R38):
Enhanced the script to include detailed checks for `gitleaks`,
`prettier`, `format`, and `ruff`. Each check now provides clear success
or failure messages and exits on failure to ensure code quality before
commits.
*
[`githooks/pre-commit-scripts/python.sh`](diffhunk://#diff-f13793c94aec64ea3a3f3f3bcc4f045eeb9d66b5d895396175dd6f49ae3302b1L1-L7):
Removed the script as its functionality is now integrated into the main
pre-commit hook script.

### Ruff checks addition:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR42-R46):
Added a new `ruff-checks` command to run both `ruff-format-check` and
`ruff-lint-check` to ensure all Ruff issues are checked.

### Dependency updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L6-R13):
Updated `ruff` to version `0.9.9`, `zizmor` to version `1.4.1`, and `uv`
required-version to `0.6.3` to keep dependencies current and secure.