Remove CodeQL Config (#180)
# Pull Request

## Description

This pull request removes the use of a custom CodeQL configuration file
and simplifies the workflow configuration for CodeQL analysis. The most
important changes are as follows:

### Removal of CodeQL custom configuration:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Deleted the query filter that excluded the `actions/unpinned-tag` query.

### Workflow simplification:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L166):
Removed the reference to the custom CodeQL configuration file
(`.github/other-configurations/codeql-config.yml`) in the `config-file`
parameter of the CodeQL setup step.