Pin GitHub Actions (#183)
# Pull Request

## Description

This pull request introduces the integration of the `pinact` tool for
workflow validation and updates some dependencies in the Python project.
The most significant changes include the addition of `pinact`
configuration and commands, updates to the pre-commit hooks to include
`pinact`, and minor dependency version updates in `pyproject.toml`.

### Integration of `pinact` for workflow validation:
*
[`.github/other-configurations/pinact.yml`](diffhunk://#diff-c2e684c61036562879b1062fb69ab67d446e19354f7a45ef92b14f611a051bc5R1-R9):
Added a new configuration file for `pinact` with schema validation and
rules to ignore specific actions.
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR111-R126):
Added new commands for running, verifying, and updating `pinact`
workflows (`pinact-run`, `pinact-check`, `pinact-update`).
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24):
Updated pre-commit hooks to include a `Pinact Checks` step that runs
`pinact-check`.

### Dependency updates:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L6-R13):
Updated `ruff` from version `0.11.6` to `0.11.7` and `uv` from version
`0.6.14` to `0.6.17`.