Update Dependabot Schedule (#192)
# Pull Request

## Description

This pull request updates the Dependabot configuration to use a
cron-based schedule instead of a daily interval for dependency updates.
The changes ensure more precise control over when updates are checked.

### Dependabot configuration updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L9-R10):
Changed the schedule for updates from a daily interval with a specific
time and timezone to a cron-based schedule (`30 7 * * *`) for both the
GitHub Actions and Python dependency groups.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L9-R10)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L24-R24)