Swap SuperLinter to Full Version (#207)
# Pull Request

## Description

This pull request includes updates to dependencies and tooling
configurations to ensure compatibility and improve functionality. The
most important changes involve updating versions for linters, hooks, and
dependencies.

### Tooling Updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L29-R29):
Updated the `super-linter` GitHub Action to use the main repository
instead of the `slim` variant, maintaining the same version (`v7.3.0`).
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Increased the minimum required version of `lefthook` from `1.11.4` to
`1.11.13`.

### Dependency Updates:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L6-R12):
Updated the `ruff` dependency from `0.11.9` to `0.11.10` and the `uv`
required version from `0.7.3` to `0.7.4`.