Remove `.editorconfig-checker.json` (#213)
# Pull Request

## Description

This pull request removes the `.editorconfig-checker.json` configuration
file, which previously contained settings for the editorconfig checker
tool.

Configuration cleanup:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Removed the configuration file, which included options such as
`Verbose`, `Debug`, and `SpacesAfterTabs`. This might indicate that the
editorconfig checker is no longer in use or its configuration is being
managed elsewhere.